--- makeMtree/chain.cpp
@@ -1,2 +1,0 @@
-#include "chain.h"
-
