+++ tools/KISA_SHA256.h
